# api/openapi.yaml
openapi: 3.1.0
info:
  title: Katalog API
  description: >
    REST API сервиса Katalog. Источник правды для контрактов:
    клиентские типы генерируются из этого файла (orval).
    Аутентификация: сессионная httpOnly cookie `katalog_session`.
    Тенант-изоляция: любой чужой или несуществующий ресурс — 404.
  version: 0.5.0
servers:
  - url: /
paths:
...
  /api/v1/admin/photos/{photoID}/block:
    post:
      operationId: adminBlockPhoto
      summary: "[admin] Скрыть фото (status=blocked, деривативы удаляются из CDN)"
      description: >
        Оригинал в S3 сохраняется. Владелец магазина получает email.
        Действие идемпотентно и пишется в аудит-лог.
      parameters:
        - name: photoID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminActionBody" }
      responses:
        "200":
          description: Фото заблокировано
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string, const: blocked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/admin/albums/{albumID}/hide:
    post:
      operationId: adminHideAlbum
      summary: "[admin] Скрыть альбом с витрины (is_hidden), email владельцу"
      parameters:
        - name: albumID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminActionBody" }
      responses:
        "204": { description: Альбом скрыт }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/admin/shops/{shopID}/suspend:
    post:
      operationId: adminSuspendShop
      summary: "[admin] Снять магазин с публикации (status=suspended), email владельцу"
      parameters:
        - $ref: "#/components/parameters/ShopID"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminActionBody" }
      responses:
        "204": { description: Магазин снят с публикации }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  parameters:
...
    ShopID:
      name: shopID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Нет или истекла сессия / неверные креды
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ресурс не найден или принадлежит другому пользователю
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: Машинный код ошибки }
        message: { type: string }
...
    AdminActionBody:
      type: object
      description: Необязательная привязка действия модератора к жалобе (аудит-лог)
      properties:
        complaint_id: { type: string, format: uuid }
        note: { type: string }
...